refactor: organize benchmarks
